requirements.txt
jinja2==3.1.2
markdown==3.5.1
pyyaml==6.0.1
pygments==2.17.2
flask==3.0.0
